AI BitURShift's result should not be unsigned https://bugs.webkit.org/show_bug.cgi?id=198689 <rdar://problem/51550063>
Reviewed by Saam Barati.

JSTests:

* stress/urshift-int32-overflow.js: Added.
(foo.):
(foo):

Source/JavaScriptCore:

Treating BitURShift's result as unsigned in the abstract interpreter incorrectly overflows it.
This breaks the DFG and FTL, since they assume that BitURShift's result is an int32 value, but
get a double constant from AI. Since the result will be converted to unsigned by UInt32ToNumber,
all we have to do is store the result as a signed int32.

* dfg/DFGAbstractInterpreterInlines.h:

Identifier: 212748@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@246332 268f45cc-cd09-0410-ab3c-d52691b4dbfc