[WHLSL] The name resolver does not deal with nativeFunctionDeclaration https://bugs.webkit.org/show_bug.cgi?id=198306
Reviewed by Saam Barati.

We currently have a crash in the nameResolver when trying to use the full standard library.
What is happening is that because we don't specify anything to do to nativeFunctionDeclarations, names in their parameters
are added to the global environment. And so as soon as we have two such parameters with the same name, the name resolver fails.

Tested by adding two native functions that share a parameter name to the standard library.

* Modules/webgpu/WHLSL/WHLSLNameResolver.cpp:
(WebCore::WHLSL::NameResolver::visit):
* Modules/webgpu/WHLSL/WHLSLNameResolver.h:


Identifier: 212916@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@246524 268f45cc-cd09-0410-ab3c-d52691b4dbfc