[iOS] Pressing key while holding Command should not insert character https://bugs.webkit.org/show_bug.cgi?id=198925 <rdar://problem/51778811>
Reviewed by Brent Fulgham.

Source/WebKit:

Do not insert a character for an unhandled key command that has a Command modifier.
For example, pressing Command + Shift + v, which is an unhandled key command (at the
time of writing) should not insert v. This matches iOS and Mac platform conventions.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _interpretKeyEvent:isCharEvent:]):

LayoutTests:

Add a test that Command + Shift + v does not insert a v as we don't expect it
to as of the time of writing. A more comprehensive test would be needed to
ensure that all unhandled key commands with Command modifiers do not insert
a character. For now, the added test seems good enough.

* fast/events/ios/command+shift+v-should-not-insert-v-expected.txt: Added.
* fast/events/ios/command+shift+v-should-not-insert-v.html: Added.

Identifier: 212956@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@246564 268f45cc-cd09-0410-ab3c-d52691b4dbfc