[GTK][WPE] Remove Flatpak sandbox https://bugs.webkit.org/show_bug.cgi?id=199416
Patch by Patrick Griffis <pgriffis@igalia.com> on 2019-07-03
Reviewed by Michael Catanzaro.

The flatpak-spawn based sandbox was only a proof of concept and
lacks flexibility for our needs so we ended up using it with very
permissive permissions and it still imposed limitations that our main
bubblewrap sandbox didn't have.

So in order to get as many applications using the sandbox as possible we
will just remove this for now and focus on bubblewrap. At some point
it may be possible to improve flatpak-spawn to fit our needs.

* SourcesGTK.txt:
* SourcesWPE.txt:
* UIProcess/Launcher/glib/FlatpakLauncher.cpp: Removed.
* UIProcess/Launcher/glib/FlatpakLauncher.h: Removed.
* UIProcess/Launcher/glib/ProcessLauncherGLib.cpp:
(WebKit::ProcessLauncher::launchProcess):

Identifier: 213365@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247096 268f45cc-cd09-0410-ab3c-d52691b4dbfc